requirements.txt
# be
fastapi
uvicorn
torch
sentence-transformers
dotenv
spacy
# fe
streamlit